email-validator/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
aiohttp>=3.9.0
dnspython>=2.6.0
beautifulsoup4>=4.12.0
